# contracts/openapi/content-service-v1.yaml
openapi: 3.1.0
info:
  title: Content Service Admin API
  version: 1.0.0
  description: Minimal Phase 1.5 admin integration contract. Development header authentication is non-production.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Deployment-relative; clients configure the environment-specific origin
paths:
...
  /api/v1/admin/sources:
    get: {operationId: listSources, summary: listSources, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], parameters: [{name: page,in: query,schema: {type: integer,default: 0}}, {name: size,in: query,schema: {type: integer,default: 20}}, {name: search,in: query,schema: {type: string}}, {name: publisher,in: query,schema: {type: string}}, {name: sourceType,in: query,schema: {$ref: '#/components/schemas/SourceType'}}, {name: reviewStatus,in: query,schema: {$ref: '#/components/schemas/ReviewStatus'}}, {name: status,in: query,schema: {$ref: '#/components/schemas/SourceStatus'}}, {name: accessedBefore,in: query,schema: {type: string,format: date}}, {name: accessedAfter,in: query,schema: {type: string,format: date}}, {name: sort,in: query,schema: {type: string}}], responses: {'200': {description: Source page,content: {application/json: {schema: {$ref: '#/components/schemas/SourcePage'}}}}}}
    post: {operationId: createSource, summary: createSource,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],requestBody: {required: true,content: {application/json: {schema: {$ref: '#/components/schemas/SourceRequest'}}}},responses: {'201': {description: Created,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}},'409': {$ref: '#/components/responses/Conflict'},'422': {$ref: '#/components/responses/Validation'}}}
  /api/v1/admin/sources/{sourceId}:
    parameters: [{$ref: '#/components/parameters/SourceId'}]
    get: {operationId: getSource, summary: getSource,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],responses: {'200': {description: Source,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}},'404': {$ref: '#/components/responses/NotFound'}}}
    put: {operationId: updateSource, summary: updateSource,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],requestBody: {required: true,content: {application/json: {schema: {$ref: '#/components/schemas/SourceRequest'}}}},responses: {'200': {description: Updated,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}},'409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/sources/{sourceId}/review:
    parameters: [{$ref: '#/components/parameters/SourceId'}]
    post: {operationId: reviewSource, summary: reviewSource,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],requestBody: {$ref: '#/components/requestBodies/VersionOnly'},responses: {'200': {description: Reviewed,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}}}
  /api/v1/admin/sources/{sourceId}/require-update:
    parameters: [{$ref: '#/components/parameters/SourceId'}]
    post: {operationId: requireSourceUpdate, summary: requireSourceUpdate,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],requestBody: {$ref: '#/components/requestBodies/VersionOnly'},responses: {'200': {description: Marked,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}}}
  /api/v1/admin/sources/{sourceId}/retire:
    parameters: [{$ref: '#/components/parameters/SourceId'}]
    post: {operationId: retireSource, summary: retireSource,security: [{developmentAdminIdentity: [],developmentAdminRoles: []}],requestBody: {required: true,content: {application/json: {schema: {$ref: '#/components/schemas/RetireSourceRequest'}}}},responses: {'200': {description: Retired,content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}},'409': {$ref: '#/components/responses/Conflict'}}}
...
components:
  securitySchemes:
    developmentAdminIdentity:
      type: apiKey
      in: header
      name: X-Admin-Identity
      description: Non-production identity header.
    developmentAdminRoles:
      type: apiKey
      in: header
      name: X-Admin-Roles
      description: Non-production comma-separated roles. Recognized roles are CONTENT_AUTHOR, CONTENT_REVIEWER, CONTENT_PUBLISHER, and ADMIN.
  parameters:
...
    SourceId: {name: sourceId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    VersionOnly: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/VersionOnly'}}}}
...
  responses:
...
    Validation: {description: Request validation failed,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    NotFound: {description: Resource not found,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Conflict: {description: Conflict or stale version,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
  schemas:
...
    SourceType: {type: string, enum: [GOVERNMENT_WEBPAGE,GOVERNMENT_DOCUMENT,LEGISLATION,PUBLIC_AUTHORITY_GUIDANCE,LICENSED_MATERIAL,INTERNAL_RESEARCH,OTHER]}
    ReviewStatus: {type: string, enum: [UNREVIEWED,REVIEWED,REQUIRES_UPDATE]}
    SourceStatus: {type: string, enum: [ACTIVE,RETIRED]}
    VersionOnly: {type: object,required: [version],properties: {version: {type: integer,format: int64,minimum: 0}}}
...
    AuditFields: {type: object,required: [id,createdAt,updatedAt,version],properties: {id: {type: string,format: uuid},createdAt: {type: string,format: date-time},updatedAt: {type: string,format: date-time},version: {type: integer,format: int64}}}
...
    SourceRequest: {type: object,required: [publisher,title,sourceType,accessedAt],properties: {publisher: {type: string},title: {type: string},url: {type: [string,'null'],format: uri},sourceType: {$ref: '#/components/schemas/SourceType'},documentVersion: {type: [string,'null']},publicationDate: {type: [string,'null'],format: date},accessedAt: {type: string,format: date},copyrightNotes: {type: [string,'null']},internalNotes: {type: [string,'null']},replacementSourceId: {type: [string,'null'],format: uuid},version: {type: integer,format: int64}}}
    Source: {allOf: [{$ref: '#/components/schemas/AuditFields'},{$ref: '#/components/schemas/SourceRequest'},{type: object,required: [reviewStatus,status],properties: {reviewStatus: {$ref: '#/components/schemas/ReviewStatus'},status: {$ref: '#/components/schemas/SourceStatus'}}}]}
    SourcePage: {type: object,required: [items,page,size,totalItems,totalPages],properties: {items: {type: array,items: {$ref: '#/components/schemas/Source'}},page: {type: integer},size: {type: integer},totalItems: {type: integer,format: int64},totalPages: {type: integer,format: int64}}}
    RetireSourceRequest: {type: object,required: [version],properties: {replacementSourceId: {type: [string,'null'],format: uuid},version: {type: integer,format: int64,minimum: 0}}}
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, timestamp, errors]
      properties:
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        errors:
          type: array
          items:
            type: object
            additionalProperties: true